move state into own file
